feat: Refactor (not finished) to implement Blueprints along the project to organize views.
