+++ .gitignore
@@ -1,0 +1,2 @@
+mindhive_env/
+__pycache__
