+++ templates/auth/login.html
@@ -23,7 +23,7 @@
 
       
 
-<form class="form-signin" action="/login" method="POST">
+<form class="form-signin" action="/auth/login" method="POST">
     
     
     <!-- <h1 class="h3 mb-3 fw-normal">Mind<strong>Hive</strong></h1> -->
@@ -40,7 +40,7 @@
     </div>
     <div class="msg">{{ msg }}</div>
     <button class="w-100 btn btn-lg btn-primary" type="submit">Inicia Sesión</button>
-    <p class="mt-3">¿No tienes cuenta? <a href="/register">Regístrate aquí</a></p>
+    <p class="mt-3">¿No tienes cuenta? <a href="/auth/signup">Regístrate aquí</a></p>
 </form>
 
 {% endblock %}
