+++ templates/auth/register.html
@@ -23,7 +23,7 @@
 
       
 
-<form class="form-signin" action="/register" method="POST">
+<form class="form-signin" action="/auth/signup" method="POST">
     
     
     <!-- <h1 class="h3 mb-3 fw-normal">Mind<strong>Hive</strong></h1> -->
@@ -59,7 +59,7 @@
     </div>
     <div class="msg">{{ msg }}</div>
     <button class="w-100 btn btn-lg btn-primary" type="submit">Registrarse</button>
-    <p class="mt-3">¿Ya tienes cuenta? <a href="/login">Inicia Sesión aquí</a></p>
+    <p class="mt-3">¿Ya tienes cuenta? <a href="/auth/login">Inicia Sesión aquí</a></p>
 </form>
 
 {% endblock %}
